refactor(lazy): move lazy config to plugins/configs dir
