+++ init.lua
@@ -6,4 +6,4 @@ require 'dtsf.disable_builtins'
 require 'dtsf.options'
 require 'dtsf.highlights'
 require 'dtsf.bindings'
-require 'dtsf.lazy'
+require 'dtsf.plugins.configs.lazy'
